Loads 21 pain studies

Cleaned data from 21 pain studies used by Maumet et al. (2016) and downloaded from Neurovault.org at http://neurovault.org/collections/1425/.

pain21()

Details

Gorgolewski KJ, Varoquaux G, Rivera G, Schwartz Y, Ghosh SS, Maumet C, Sochat VV, Nichols TE, Poldrack RA, Poline J-B, Yarkoni T and Margulies DS (2015) NeuroVault.org: a web-based repository for collecting and sharing unthresholded statistical maps of the brain. Front. Neuroinform. 9:8. doi: 10.3389/fninf.2015.00008 Maumet, Camille, Tibor Auer, Alexander Bowring, Gang Chen, Samir Das, Guillaume Flandin, Satrajit Ghosh, et al. “Sharing Brain Mapping Statistical Results with the Neuroimaging Data Model.” Scientific Data 3 (December 6, 2016). https://doi.org/10.1038/sdata.2016.102.

Examples

pain21()
#> $data
#>      study  n
#> 1  pain_01 25
#> 2  pain_02 25
#> 3  pain_03 20
#> 4  pain_04 20
#> 5  pain_05  9
#> 6  pain_06  9
#> 7  pain_07  9
#> 8  pain_08 12
#> 9  pain_09 12
#> 10 pain_10 12
#> 11 pain_11 12
#> 12 pain_12 13
#> 13 pain_13 32
#> 14 pain_14 24
#> 15 pain_15 14
#> 16 pain_16 14
#> 17 pain_17 12
#> 18 pain_18 12
#> 19 pain_19 16
#> 20 pain_20 16
#> 21 pain_21 16
#>                                                                   images
#> 1  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_01.nii.gz
#> 2  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_02.nii.gz
#> 3  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_03.nii.gz
#> 4  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_04.nii.gz
#> 5  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_05.nii.gz
#> 6  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_06.nii.gz
#> 7  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_07.nii.gz
#> 8  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_08.nii.gz
#> 9  /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_09.nii.gz
#> 10 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_10.nii.gz
#> 11 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_11.nii.gz
#> 12 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_12.nii.gz
#> 13 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_13.nii.gz
#> 14 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_14.nii.gz
#> 15 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_15.nii.gz
#> 16 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_16.nii.gz
#> 17 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_17.nii.gz
#> 18 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_18.nii.gz
#> 19 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_19.nii.gz
#> 20 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_20.nii.gz
#> 21 /home/runner/work/_temp/Library/pain21/pain21/contrast_pain_21.nii.gz
#>                                                           varimages
#> 1  /home/runner/work/_temp/Library/pain21/pain21/var_pain_01.nii.gz
#> 2  /home/runner/work/_temp/Library/pain21/pain21/var_pain_02.nii.gz
#> 3  /home/runner/work/_temp/Library/pain21/pain21/var_pain_03.nii.gz
#> 4  /home/runner/work/_temp/Library/pain21/pain21/var_pain_04.nii.gz
#> 5  /home/runner/work/_temp/Library/pain21/pain21/var_pain_05.nii.gz
#> 6  /home/runner/work/_temp/Library/pain21/pain21/var_pain_06.nii.gz
#> 7  /home/runner/work/_temp/Library/pain21/pain21/var_pain_07.nii.gz
#> 8  /home/runner/work/_temp/Library/pain21/pain21/var_pain_08.nii.gz
#> 9  /home/runner/work/_temp/Library/pain21/pain21/var_pain_09.nii.gz
#> 10 /home/runner/work/_temp/Library/pain21/pain21/var_pain_10.nii.gz
#> 11 /home/runner/work/_temp/Library/pain21/pain21/var_pain_11.nii.gz
#> 12 /home/runner/work/_temp/Library/pain21/pain21/var_pain_12.nii.gz
#> 13 /home/runner/work/_temp/Library/pain21/pain21/var_pain_13.nii.gz
#> 14 /home/runner/work/_temp/Library/pain21/pain21/var_pain_14.nii.gz
#> 15 /home/runner/work/_temp/Library/pain21/pain21/var_pain_15.nii.gz
#> 16 /home/runner/work/_temp/Library/pain21/pain21/var_pain_16.nii.gz
#> 17 /home/runner/work/_temp/Library/pain21/pain21/var_pain_17.nii.gz
#> 18 /home/runner/work/_temp/Library/pain21/pain21/var_pain_18.nii.gz
#> 19 /home/runner/work/_temp/Library/pain21/pain21/var_pain_19.nii.gz
#> 20 /home/runner/work/_temp/Library/pain21/pain21/var_pain_20.nii.gz
#> 21 /home/runner/work/_temp/Library/pain21/pain21/var_pain_21.nii.gz
#> 
#> $mask
#> [1] "/home/runner/work/_temp/Library/pain21/pain21/mask.nii.gz"
#> 
#> $template
#> [1] "/home/runner/work/_temp/Library/pain21/pain21/MNI152_T1_2mm_brain.nii.gz"
#>